SQL #8/9: DML & DDL
DML (Data Manipulation Language) refers to the subset of SQL commands used to interact with and manipulate data stored in a database. DML commands are essential for performing operations like inserting, updating, deleting, and querying data.
DDL (Data Definition Language) consists of SQL commands that define and modify the structure of database objects such as tables, indexes, and schemas. DDL commands are used to create, alter, and delete these objects.